+++ pruebas/01-reading-list/Davidvn95/reading-lest-davidvn95/package.json
@@ -7,7 +7,8 @@
         "dev": "vite --host 0.0.0.0",
         "build": "vite build",
         "lint": "eslint src --ext js,jsx --report-unused-disable-directives --max-warnings 0",
-        "preview": "vite preview"
+        "preview": "vite preview",
+        "test": "vitest"
     },
     "dependencies": {
         "@fortawesome/free-regular-svg-icons": "^6.4.0",
@@ -20,13 +21,19 @@
         "react-router-dom": "^6.14.1"
     },
     "devDependencies": {
+        "@testing-library/jest-dom": "^5.17.0",
+        "@testing-library/react": "^14.0.0",
+        "@testing-library/user-event": "^14.4.3",
         "@types/react": "^18.2.14",
         "@types/react-dom": "^18.2.6",
         "@vitejs/plugin-react-swc": "^3.3.2",
+        "@vitest/ui": "^0.33.0",
         "eslint": "^8.44.0",
         "eslint-plugin-react": "^7.32.2",
         "eslint-plugin-react-hooks": "^4.6.0",
         "eslint-plugin-react-refresh": "^0.4.1",
-        "vite": "^4.4.0"
+        "jsdom": "^22.1.0",
+        "vite": "^4.4.0",
+        "vitest": "^0.33.0"
     }
 }
