+++ pruebas/01-reading-list/Davidvn95/reading-lest-davidvn95/src/components/NavBar/NavBar.css
@@ -76,102 +76,6 @@
     outline: none;
 }
 
-.menu {
-  background-color: transparent;
-  border: none;
-  cursor: pointer;
-  display: flex;
-  padding: 0;
-  z-index: 10;
-}
-.line {
-  fill: none;
-  stroke: var(--color-menu);
-  stroke-width: 4;
-  transition: stroke-dasharray 600ms cubic-bezier(0.4, 0, 0.2, 1),
-    stroke-dashoffset 600ms cubic-bezier(0.4, 0, 0.2, 1);
-}
-.line1 {
-  stroke-dasharray: 60 207;
-  stroke-width: 4;
-}
-.line2 {
-  stroke-dasharray: 60 60;
-  stroke-width: 4;
-}
-.line3 {
-  stroke-dasharray: 60 207;
-  stroke-width: 4;
-}
-.opened .line1 {
-  stroke-dasharray: 90 207;
-  stroke-dashoffset: -134;
-  stroke-width: 4;
-}
-.opened .line2 {
-  stroke-dasharray: 1 60;
-  stroke-dashoffset: -30;
-  stroke-width: 4;
-}
-.opened .line3 {
-  stroke-dasharray: 90 207;
-  stroke-dashoffset: -134;
-  stroke-width: 4;
-}
-.mobile{
-    display: flex;
-    flex-direction: column;
-    position: relative;
-}
-
-.mobile-nav{
-  position: fixed;
-  right: 0;
-  top: 0;
-  max-width: 75%;
-  min-height: 78vh;
-  border: 1px solid gray;
-  border-radius: 1rem 0 0 1rem;
-  gap: 2rem;
-  display: flex;
-  flex-direction: column;
-  padding: 1rem 2rem;
-  background-color: var(--color-background);
-}
-
-.mobile-links{
-  display: flex;
-  flex-direction: column;
-}
-
-.mobile-links > a{
-  font-size: x-large;
-  text-decoration: none;
-  font-weight: 700;
-  color: var(--color-font);
-}
-
-
-.mobile-search{
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
-
-.mobile-search > input{
-    background-color: transparent;
-    border: 2px solid rgb(110, 84, 18);
-    border-radius: 1rem;
-    height: 1.5rem;
-    width: 16rem;
-    font-weight: 700;
-    padding: 0 1rem;
-}
-
-.mobile-search > span{
-  font-size: x-large;
-  font-weight: 700;
-}
 
 @media screen and (max-width: 768px) {
     .navbar{
