+++ pruebas/01-reading-list/Davidvn95/reading-lest-davidvn95/src/data/StorgeListener/StorageListener.jsx
@@ -1,5 +1,5 @@
-import { store } from "../redux/store";
-import { syncsTabs } from "../redux/booksSlice";
+import { store } from "../../redux/store";
+import { syncsTabs } from "../../redux/booksSlice";
 import { useEffect } from "react";
 
 export default function StorageListener() {
