+++ pruebas/01-reading-list/Davidvn95/reading-lest-davidvn95/src/main.jsx
@@ -4,7 +4,7 @@ import { BrowserRouter } from "react-router-dom";
 import { Provider } from "react-redux";
 import { store } from "./redux/store";
 import App from "./App.jsx";
-import StorageListener from "./components/storageListener";
+import StorageListener from "./data/StorgeListener/storageListener";
 import "./index.css";
 
 ReactDOM.createRoot(document.getElementById("root")).render(
