+++ pruebas/01-reading-list/Davidvn95/reading-lest-davidvn95/vite.config.js
@@ -1,7 +1,17 @@
-import { defineConfig } from 'vite'
-import react from '@vitejs/plugin-react-swc'
+/// <reference types="vitest" />
+/// <reference types="vite/client" />
+
+import { defineConfig } from "vite";
+import react from "@vitejs/plugin-react-swc";
+// import swcOptions from "./.swcrc";
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [react()],
-})
+    plugins: [react()],
+    test: {
+        globals: true,
+        environment: "jsdom",
+        // transform: { "^.+.jsx?$": "swc" },
+        // swcOptions,
+    },
+});
